requirements.txt
# Core Dependencies
requests>=2.31,<3 --only-binary=:all:
psycopg2-binary>=2.9.10,<3 --only-binary=:all:
python-dotenv>=1.0,<2 --only-binary=:all:

# NLP & Text Processing
nltk>=3.8,<4 --only-binary=:all:
beautifulsoup4>=4.12,<5 --only-binary=:all:
html5lib>=1.1,<2 --only-binary=:all:
lxml>=5.1,<6 --only-binary=:all:

# Machine Learning & Data
scikit-learn>=1.4,<2 --only-binary=:all:
numpy>=2.2,<3 --only-binary=:all:
pandas>=2.2,<3 --only-binary=:all:

# Sentiment Analysis (VADER)
vaderSentiment>=3.3,<4 --only-binary=:all:

# Visualization
wordcloud>=1.9,<2 --only-binary=:all:
matplotlib>=3.8,<4 --only-binary=:all:
seaborn>=0.13,<1 --only-binary=:all:

# Misc
tqdm>=4.66,<5 --only-binary=:all:
